Added the first version of requiredVars
This scripts parses the xlm/dreq tree to find the requested variables
on each MIP tables.  It then creates an Excel spreadsheet listing
the variable name, its official title and units, and a definition if
one is found within the user supplied definition file.  A list of MIPS
requesting this table is also listed.  Each MIP table has its own sheet
within the worksheet.

The idea of this script is to provide the output to the scientists to
get a full list of definitions.