Some cleanup in createOutputSpecs
NOTE: When using a dictionary, do not check if an object is "in" the
dictionary "keys" (i.e., "x in dict.keys()").  One should check if the
object is in the dictionary proper (i.e., "x in dict").  The "keys" 
function returns a list, and checking if an object is in a list is
linear in the length of the list.  Checking directly if an object is in
a dictionary is constant in time because the dictionary uses a hash
table for lookup.