Allow for multiplication/division operations across common dimensions.
This feature makes it possible to extend the dimensions of an array
through simple multiplication or division across common dimensions.  The
idea is that two arrays, for example A['x'] and B['y'], depending on
dimensions 'x' and 'y' can be multiplied together unambiguously even
though their dimensions are different.  The resulting array shoud depend
upon both dimensions, such that:

    A['x'] * B['y'] = C['x','y']

or, essentially, the outer product.  This is true for both
multiplication and division, but cannot hold for addition/subtraction.
In the case of exponentiation, the exponent must be scalar, so this
"expansive" definition also does not hold.  In the case of the modulo
operation, the modulo can be written as:

	A % B = A - B * (A // B)

where // represents "floor division".  This means that if the
"expansive" definition of multiplication and division is used, the
dimensions of this equation become:

	A['x'] * B['y'] = A['x'] - B['y'] * (A['x'] // B['y'])
	                = A['x'] - C['y', 'x']

and this cannot be represented in the "expansive" form, because
addition/subtraction cannot be represented in an "expansive" form.
Hence, the modulo operation requires a special implementation, too.