Testing NC3-style output files with accepted datatypes
NC3 only accepts int8/16/32, float32/64, and 'S1' dtypes.  Should error
out with useful error message if another type is encountered, instead of
trying to convert (if not obvious how to do so).  

This commit changes the tests to only use NC3-compliant dtypes so that
these errors are not encountered when testing NC3 output.